Fix references to external dependencies
Add references to the locations of the Boost, Crypto++, and Curl headers and libraries.  Add a property sheet which defines user macros that point to the roots of those items, as well as the include and lib directories.  The property sheet provides a single location to be modified to configure the external dependencies when moving to a new machine.