Buildfixes for newer libraries
A few minor fixes are needed to compile with the VC++ 2013 headers and Boost 1.50 or later.

Include <algorithm> so std::min/max is available.

Get rid of unused 'int round(float)' which conflicts with 'float round(float)' in <cmath>.

boost::TIME_UTC is now boost::TIME_UTC_, as of 1.50.  Use an #if BOOST_VERSION to choose the proper name.