Fix sending GPT footer with large volumes
Volumes with size greater than 2 TiB have GPT header and footer stored for
convenience. We need to substract the GPT footer when restoring otherwise
the volume might not fit on the partition.

Get the size of the volume without GPT footer from the GPT partition data
and restore the correctly sized volume.